refactor graph embeddings to a separate module called embeding.py
